requirements.txt
amqp==5.0.2
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.4
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
conda==4.9.2
conda-package-handling==1.7.0
configparser==5.0.1
crayons==0.4.0
cryptography==2.9.2
Django==3.1.2
django-appconf==1.0.4
django-celery-inspect==0.2.1
django-ckeditor==6.0.0
django-compressor==2.4
django-debug-toolbar==3.1.1
django-http-monitor==0.5.5
django-js-asset==1.2.2
django-libsass==0.8
djangorestframework==3.12.2
gTTS==2.2.1
idna==2.8
kombu==5.0.2
libsass==0.20.1
menuinst==1.4.16
mysqlclient==2.0.1
nose==1.3.7
playsound==1.2.2
prompt-toolkit==3.0.8
pycosat==0.6.3
pycparser==2.20
pygal==2.4.0
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2021.1
pywin32==227
rcssmin==1.0.6
redis==3.5.3
requests==2.22.0
rjsmin==1.1.0
ruamel-yaml==0.15.87
selenium==3.141.0
six==1.14.0
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tqdm==4.46.0
urllib3==1.25.8
vboxapi==1.0
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.2.2
wikipedia==1.4.0
win-inet-pton==1.1.0
wincertstore==0.2
youtube-search==1.1.1
